# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VeriFactu SaaS API - Sistema de Facturación AEAT
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List user's organizations
      responses:
        "200":
          description: Organizations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationBody"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{id}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationBody"
      responses:
        "200":
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"

  # Taxpayers
...
  /taxpayers/{taxpayerId}/products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products and services
      parameters:
        - name: taxpayerId
          in: path
          required: true
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product or service
      parameters:
        - name: taxpayerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Series
  /taxpayers/{taxpayerId}/series:
    get:
      operationId: listSeries
      tags: [series]
      summary: List invoice series
      parameters:
        - name: taxpayerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Series list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvoiceSeries"
    post:
      operationId: createSeries
      tags: [series]
      summary: Create invoice series
      parameters:
        - name: taxpayerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSeriesBody"
      responses:
        "201":
          description: Series created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceSeries"
  /series/{id}:
    patch:
      operationId: updateSeries
      tags: [series]
      summary: Update series
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSeriesBody"
      responses:
        "200":
          description: Series updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceSeries"

  # Invoices
...
  /organizations/{orgId}/gestoria/overview:
    get:
      operationId: getGestoriaOverview
      tags: [organizations]
      summary: Get gestoria overview across all managed taxpayers
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gestoria overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GestoriaOverview"
  /organizations/{orgId}/gestoria/incidents:
    get:
      operationId: listGestoriaIncidents
      tags: [organizations]
      summary: List AEAT incidents across all managed taxpayers
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Incidents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AeatIncident"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Organization:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [autonomo, empresa, gestoria]
        nif:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        country:
          type: string
          default: "ES"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        role:
          type: string
          enum: [owner, admin, facturador, asesor, readonly]
        taxpayerCount:
          type: integer
      required:
        - id
        - name
        - type
        - country
        - createdAt
        - updatedAt
        - role
        - taxpayerCount
    CreateOrganizationBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [autonomo, empresa, gestoria]
        nif:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
      required:
        - name
        - type
    UpdateOrganizationBody:
      type: object
      properties:
        name:
          type: string
        nif:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        province:
          type: ["string", "null"]

    # Taxpayer
...
    Product:
      type: object
      properties:
        id:
          type: integer
        taxpayerId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        unitPrice:
          type: number
        vatRate:
          type: number
        unit:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - taxpayerId
        - name
        - unitPrice
        - vatRate
        - isActive
        - createdAt
        - updatedAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        unitPrice:
          type: number
        vatRate:
          type: number
        unit:
          type: ["string", "null"]
      required:
        - name
        - unitPrice
        - vatRate
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        unitPrice:
          type: number
        vatRate:
          type: number
        unit:
          type: ["string", "null"]

    # Series
    InvoiceSeries:
      type: object
      properties:
        id:
          type: integer
        taxpayerId:
          type: integer
        prefix:
          type: string
        name:
          type: string
        currentNumber:
          type: integer
        year:
          type: integer
        isDefault:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - taxpayerId
        - prefix
        - name
        - currentNumber
        - year
        - isDefault
        - isActive
        - createdAt
    CreateSeriesBody:
      type: object
      properties:
        prefix:
          type: string
        name:
          type: string
        startNumber:
          type: integer
        year:
          type: integer
        isDefault:
          type: boolean
      required:
        - prefix
        - name
    UpdateSeriesBody:
      type: object
      properties:
        name:
          type: string
        isDefault:
          type: boolean
        isActive:
          type: boolean

    # Invoice
...
    GestoriaOverview:
      type: object
      properties:
        totalTaxpayers:
          type: integer
        activeTaxpayers:
          type: integer
        pendingSetup:
          type: integer
        pendingSubmissions:
          type: integer
        pendingIncidents:
          type: integer
        taxpayerSummaries:
          type: array
          items:
            $ref: "#/components/schemas/TaxpayerSummary"
      required:
        - totalTaxpayers
        - activeTaxpayers
        - pendingSetup
        - pendingSubmissions
        - pendingIncidents
        - taxpayerSummaries
    TaxpayerSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nif:
          type: string
        pendingInvoices:
          type: integer
        pendingSubmissions:
          type: integer
        hasIncidents:
          type: boolean
        lastActivity:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - name
        - nif
        - pendingInvoices
        - pendingSubmissions
        - hasIncidents
    AeatIncident:
      type: object
      properties:
        id:
          type: integer
        taxpayerId:
          type: integer
        taxpayerName:
          type: string
        invoiceId:
          type: integer
        invoiceNumber:
          type: ["string", "null"]
        errorCode:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - taxpayerId
        - taxpayerName
        - invoiceId
        - status
        - createdAt